# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DRIP. API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /similar/{scanId}:
    get:
      operationId: getSimilarOutfits
      tags: [style]
      summary: Get similar outfit inspiration
      security:
        - bearerAuth: []
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar outfits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimilarOutfit"
  /recommendations/{scanId}:
    get:
      operationId: getRecommendations
      tags: [style]
      summary: Get upgrade recommendations
      security:
        - bearerAuth: []
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Upgrade recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UpgradeRecommendation"
  /style-dna:
    get:
      operationId: getStyleDna
      tags: [style]
      summary: Get user Style DNA profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Style DNA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleDnaResponse"
  /style-dna/report:
    get:
      operationId: getStyleReport
      tags: [style]
      summary: Get monthly style report
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Style report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleReportResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpgradeRecommendation:
      type: object
      properties:
        swap:
          type: string
        with:
          type: string
        predictedScoreIncrease:
          type: number
        searchQuery:
          type: string
        priceRange:
          type: string
      required:
        - swap
        - with
        - predictedScoreIncrease
        - searchQuery
        - priceRange
...
    ScanSummary:
      type: object
      properties:
        id:
          type: integer
        imagePath:
          type: string
        overallScore:
          type: number
        styleArchetype:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - imagePath
        - overallScore
        - styleArchetype
        - createdAt
...
    SimilarOutfit:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        sourceUrl:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        styleTag:
          type: ["string", "null"]
      required:
        - id
        - imageUrl
    StyleDnaResponse:
      type: object
      properties:
        locked:
          type: boolean
        scansNeeded:
          type: integer
        dominantArchetype:
          type: ["string", "null"]
        archetypePercentage:
          type: ["number", "null"]
        topColors:
          type: array
          items:
            type: string
        avgScores:
          type: ["object", "null"]
        totalOutfitsScanned:
          type: integer
        bestOutfit:
          $ref: "#/components/schemas/ScanSummary"
        worstOutfit:
          $ref: "#/components/schemas/ScanSummary"
        styleEvolution:
          type: array
          items:
            type: object
            properties:
              archetype:
                type: string
              date:
                type: string
                format: date-time
              score:
                type: number
            required:
              - archetype
              - date
              - score
        scoreTrend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              score:
                type: number
            required:
              - date
              - score
      required:
        - locked
        - scansNeeded
        - totalOutfitsScanned
    StyleReportResponse:
      type: object
      properties:
        available:
          type: boolean
        totalOutfits:
          type: integer
        avgScore:
          type: number
        scoreTrend:
          type: string
          enum: [up, down, stable]
        mostWornStyle:
          type: ["string", "null"]
        bestOutfit:
          $ref: "#/components/schemas/ScanSummary"
        topColorPalette:
          type: array
          items:
            type: string
        styleInOneWord:
          type: ["string", "null"]
      required:
        - available
        - totalOutfits
        - avgScore
        - scoreTrend
...