# lib/api-zod/src/generated/types/similarOutfit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DRIP. API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimilarOutfit {
  id: number;
  imageUrl: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  styleTag?: string | null;
}